test: finaliza teste para viewmodel
