--- Testes-Unitarios/Chuck/ChuckTests/ViewModelTests.swift
@@ -1,46 +1,0 @@
-//
-//  ViewModelTests.swift
-//  ChuckTests
-//
-//  Created by Jenifer Rocha on 12/08/25.
-//
-
-import XCTest
-@testable import Chuck
-
-final class ViewModelTests: XCTestCase {
-    
-    var viewModel: HomeViewModel!
-    var mockService: MockHomeService!
-
-    override func setUpWithError() throws {
-        mockService = MockHomeService()
-        viewModel = HomeViewModel(service: mockService)
-    }
-
-    override func tearDownWithError() throws {
-        mockService = nil
-        viewModel = nil
-    }
-
-    func testFetchRequestSuccess() {
-        
-        var list: [String] = ["category1", "category2"]
-        mockService.result = .success(list)
-        
-        viewModel.fetchRequest()
-        
-        XCTAssertEqual(viewModel.numberOfRowsInSection, list.count)
-        XCTAssertEqual(viewModel.loadCurrentCategory(indexPath: IndexPath(row: 0, section: 0)), list[0])
-
-    }
-}
-
-class MockHomeService: HomeServiceProtocol {
-    var result: Result<[String], Error> = .success([])
-    func getHome(completion: @escaping (Result<[String], any Error>) -> Void) {
-        completion(result)
-    }
-    
-    
-}
